Adds a background
Backgrounds are quads. Remember to keep the original images's aspect
ratio in the scale. In this kind of ortographic camera game, you can
send the background down in the Y axis so the player isn't buried in it.
The shader is the "Unlit texture" so lighting doesn't affect it, since
it's a simple background.

http://unity3d.com/learn/tutorials/projects/space-shooter/adding-a-background?playlist=17147