Lesson 14: adding a score text, setting score
Interesting about how GUIText elements use the viewport rather than the
scene, and thus they don't use pixels for positioning nor can they be
seen in the scene view (had to switch to Game to see it). In Unity 5,
instead of being first class elements, so they have to be bound to an
empty game object.

Also learned about how to get a reference to the game controller from a
hazard to be able to update the former's score.

http://unity3d.com/learn/tutorials/projects/space-shooter/counting-points?playlist=17147