Lesson 17f: enemy prefab
Turns the enemy into a prefab (protip: if the element turns blue in the
hierarchy); adds it to the hazards array. Interesting note on
distributions: it's random so it could be crazy, but it *tends* to add 1
enemy per 3 asteroids; could add more enemies to the array if we want to
influence the frequency.

1:32 to 1:35 of
http://unity3d.com/learn/tutorials/modules/beginner/live-training-archive/extending-space-shooter?playlist=17147