Lesson 18a: unity remote, device input
Using the magic of unity remote, the accelerometer and "Fire1"'s default
binding to any touch event to stream the game physically on my iPhone!
Interesting how we need to "calibrate" to the device if we want to use
the accelerometer comfortably, rather than just use the default 3D zero
point that expects the phone to be parallel to the ground.

Made a bunch of tweaks to assets, too, so they'd render correctly on
iPhone: using truecolor for sprites, compressing the background a bit.

Had to change the build settings and also the player settings to build
to iOS and use unity remote, respectively.

First 30 mins of
http://unity3d.com/learn/tutorials/modules/beginner/live-training-archive/space-shooter-to-mobile?playlist=17147